+++ src/main.rs
@@ -206,7 +206,7 @@ async fn main() -> std::io::Result<()> {
             .service(activitypub::replies_collection)
             .service(activitypub::emoji_view)
             .service(activitypub::tag_view)
-            .service(activitypub::apresolver_view)
+            .service(activitypub::apgateway_view)
             .service(atom_scope())
             .service(nodeinfo::get_nodeinfo_jrd)
             .service(nodeinfo::get_nodeinfo_2_0)